Refactor: enhance AuthDioHelper and DioFactory for dynamic token resolution
The issue is that the unauthenticated Dio instance will NEVER have authorization headers, even if the user logs in later, because the authorization logic is embedded in the Dio creation, not in dynamic interceptors. With this commit it is made dynamic to support interceptors better so when API call has been made the correct token is given in the header.

Also Clean-up and dart formatting